Stage

+++ lib/feature/home/home_details/notstart/data/model/qadim_show_action_model.dart
@@ -46,7 +46,7 @@ class Auction {
   final int? auctionDurationMinutes;
 
   @JsonKey(name: 'auction_start_rate')
-  final int auctionStartRate;
+  final double auctionStartRate;
 
   @JsonKey(name: 'product_sku')
   final String productSku;
